migrate service.compiler to service.search
begin work to strip model.compiler down to only core compiler essentials for future rewrite.
IConceptPreflightService deleted, superceded by IPreflightResourceReader.
removed IPanelConverter, now PanelConverter is first class model type and directly reliable.
PreflightConceptReader superceded by PreflightResourceReader and delete.
