+++ src/server/Model/Search/IDemographicQueryService.cs
@@ -3,19 +3,16 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-using System.Collections.Generic;
+using System;
 using System.Threading.Tasks;
+using Model.Compiler;
 
-namespace Model.Compiler
+namespace Model.Search
 {
-    public interface IPreflightResourceReader : IPreflightConceptReader
+    public interface IDemographicQueryService
     {
-        Task<PreflightResources> GetAsync(ResourceRefs refs);
-    }
-
-    public interface IPreflightConceptReader
-    {
-        Task<PreflightConcepts> GetAsync(ConceptRef @ref);
-        Task<PreflightConcepts> GetAsync(HashSet<ConceptRef> ids);
+        Task<DemographicQuery> GetDemographicQueryAsync();
+        Task<CompilerValidationContext<DemographicCompilerContext>> GetDemographicQueryCompilerContext(QueryRef queryRef);
+        Task<DemographicQuery> UpdateDemographicQueryAsync(DemographicQuery query);
     }
 }
