+++ src/server/Services/Admin/AdminConceptService.cs
@@ -15,11 +15,10 @@ using System.Data.SqlClient;
 using System.Data;
 using Dapper;
 using Services.Tables;
-using Services.Compiler;
+using Services.Search;
 using System.Linq;
 using Model.Authorization;
 using Services.Extensions;
-using Services.Search;
 
 namespace Services.Admin
 {
