+++ src/server/Services/Cohort/CtePatientCohortService.cs
@@ -3,25 +3,15 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-using System;
+using System.Collections.Generic;
+using System.Data.SqlClient;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Data;
-using System.Data.SqlClient;
-using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Logging;
-using Model.Options;
-using Model.Compiler;
+using Microsoft.Extensions.Options;
 using Model.Cohort;
-using Services.Compiler;
-using Services.Authorization;
-using Services.Extensions;
-using Services.Cohort;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Security.Claims;
-using Model.Authorization;
+using Model.Compiler;
+using Model.Options;
 
 namespace Services.Cohort
 {
