+++ src/server/Services/Cohort/DatasetService.cs
@@ -4,21 +4,19 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
+using System.Collections.Generic;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Data.SqlClient;
-using System.Collections.Generic;
-using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Logging;
-using Services.Authorization;
-using Services.Compiler;
-using Model.Options;
+using Microsoft.Extensions.Options;
+using Model.Anonymization;
+using Model.Authorization;
 using Model.Cohort;
 using Model.Compiler;
-using Model.Anonymization;
+using Model.Options;
 using Services.Extensions;
-using Model.Authorization;
 
 namespace Services.Cohort
 {
