+++ src/server/Services/Extensions/SqlServerExtensions.cs
@@ -4,11 +4,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
-using System.Linq;
 using System.Collections.Generic;
 using System.Data.SqlClient;
+using System.Linq;
 using Model.Compiler;
-using Services.Compiler;
 using Model.Validation;
 
 namespace Services.Extensions
