+++ src/server/Services/Search/DemographicQueryRecord.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Model.Compiler;
 
-namespace Services.Compiler
+namespace Services.Search
 {
     class DemographicQueryRecord
     {
