+++ src/server/Services/Search/QueryPreflightCheckResultRecord.cs
@@ -5,7 +5,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
 
-namespace Services.Compiler
+namespace Services.Search
 {
     class QueryPreflightCheckResultRecord
     {
