+++ src/server/Tests/SearchTermTableTests.cs
@@ -4,11 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
-using System.Linq;
-using System.Collections.Generic;
-using Xunit;
-using Services.Compiler;
 using Services.Tables;
+using Xunit;
 
 namespace Tests
 {
